Add SAML 2 federation support
Adds support for configuring Keystone with mod_mellon
to support SAML 2 IdPs. This is an extension of the
existing OpenID support and uses a lot of the same
mechanisms.

This also renames the oidc attribute mapping variables
to make them generic, as all identity providers will
use the same mechanism for registering mappings.

Change-Id: I69869be87856b3fe9cb76a3214703f61af67ee22